Fix the sign-compare warnings whose signed side cannot be negative
Twenty of the thirty-three -Wsign-compare warnings are a signed loop
counter or a strlen result measured against a sizeof. In each of these
the signed operand is provably non-negative at the comparison - i counts
up from 0, len comes from strlen, q_shared.c's ret is already guarded by
ret > -1, and sv_user.c's c is the remainder of a download whose offset
the server has already rejected unless 0 <= offset < size. Making the
conversion explicit changes no behaviour.

Two of them were not casts. The Z_FreeGame guard byte is a bit pattern,
so it is read and written through uint32 rather than int - the write at
common.c:2024 moves with the two reads so the three agree. And
files.c:1076 compared (int)fread(...) against a uint32 dirlen; dropping
the cast compares the size_t that fread returns, which is also one fewer
narrowing conversion for a 64-bit MSVC build.

GCC drops from 57/55 warnings to 37/35, Clang from 53 to 34. Thirteen
sign-compare warnings remain and are deliberately untouched: those are
the ones where the signed side can reach a negative value, so the
conversion is load-bearing and each needs its own decision.

+++ game/q_shared.c
@@ -1579,7 +1579,7 @@ int Q_vsnprintf (char *buff, size_t len, const char *fmt, va_list va)
 	int ret;
 
 	ret = vsnprintf (buff, len, fmt, va);
-	if (ret > -1 && ret < len)
+	if (ret > -1 && (size_t)ret < len)
 		return ret;
 
 	return -1;
